Moved safe snapshot button & autoupdate snapshot lsit
